Fix bad implicit coercion to double

+++ mypyc/lib-rt/int_ops.c
@@ -294,7 +294,7 @@ PyObject *CPyLong_FromStr(PyObject *o) {
 }
 
 CPyTagged CPyTagged_FromFloat(double f) {
-    if (f < (CPY_TAGGED_MAX + 1.0) && f > (CPY_TAGGED_MIN - 1.0)) {
+    if (f < ((double)CPY_TAGGED_MAX + 1.0) && f > (CPY_TAGGED_MIN - 1.0)) {
         return (Py_ssize_t)f << 1;
     }
     PyObject *o = PyLong_FromDouble(f);
